src/pyproject.toml
[tool.poetry]
name = "social_bridge"
version = "0.1.0"
description = ""
authors = ["rolandsobczak <rolandsobczak@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = {extras = ["standard"], version = "^0.22.0"}
fastapi = "^0.96.0"
sqlalchemy = "^2.0.15"
alembic = "^1.11.1"
sqlalchemy-utils = "^0.41.1"
httpx = "^0.24.1"
psycopg2-binary = "^2.9.6"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
itsdangerous = "^2.1.2"
jinja2 = "^3.1.2"


[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
pylint = "^2.17.4"
black = "^23.3.0"
flake8 = "^6.0.0"
isort = "^5.12.0"
factory-boy = "^3.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
